% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_pcds.R
\name{find_nearest_pcds}
\alias{find_nearest_pcds}
\title{Order climate stations (PCDs) by their distances from a given coordinate.}
\usage{
find_nearest_pcds(lat = -22.410259, long = -47.560156)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{long}{Longitude in decimals}
}
\value{
A data frame ordered by the distances to the PCDs.
}
\description{
Order climate stations (PCDs) by their distances from a given coordinate.
}

