% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate}
\alias{get_climate}
\title{Scrape (harvest from the web) climate data from stations in Brazil.}
\usage{
get_climate(station_id = 31973, start_date = "2005/01/01",
  end_date = "2005/02/02")
}
\arguments{
\item{station_id}{A numeric vector with the station id}

\item{start_date}{Start date}

\item{end_date}{End date (Maximum of one year between start and end dates)}
}
\value{
A data frame containing the climate data and attributes.
}
\description{
Scrape (harvest from the web) climate data from stations in Brazil.
}

