% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get the geographical coordinates of a climate station (PCD)}
\usage{
get_coordinates(station_id = 31973)
}
\arguments{
\item{station_id}{A numeric vector with the station id}
}
\value{
A named numeric vector
}
\description{
Get the geographical coordinates of a climate station (PCD)
}

