% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distances.R
\name{get_distances}
\alias{get_distances}
\title{Get the distance in kilometers between two points.}
\usage{
get_distances(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1}{Latitude in decimals}

\item{long1}{Longitude in decimals}

\item{lat2}{Latitude in decimals}

\item{long2}{Longitude in decimals}
}
\value{
A numeric vector with the distance in kilometers.
}
\description{
Get the distance in kilometers between two points.
}

