% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inmet.R
\name{get_inmet_data}
\alias{get_inmet_data}
\title{Get climate data from INMET}
\usage{
get_inmet_data(station_id = 83726, start_date = "01/01/1901",
  end_date = "13/04/2016", monthly = TRUE, username, password)
}
\arguments{
\item{station_id}{The ID of the station}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{monthly}{Daily or Monthy data?}

\item{username}{INMET username}

\item{password}{INMET password}
}
\description{
Get climate data from INMET
}

