% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcds.R
\name{get_pcds}
\alias{get_pcds}
\title{List all available climate stations (PCDs)}
\usage{
get_pcds(states = "all")
}
\arguments{
\item{states}{A character vector with abbreviated Brazilian states names or "all" for all PCDs.}
}
\value{
A data frame.
}
\description{
List all available climate stations (PCDs)
}

