% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeframe.R
\name{get_timeframe}
\alias{get_timeframe}
\title{Get the span of the data of a climate station (PCD).}
\usage{
get_timeframe(station_id = 31973)
}
\arguments{
\item{station_id}{A numeric vector with the station id}
}
\value{
A character vector
}
\description{
Get the span of the data of a climate station (PCD).
}

