% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{inpe_station_data}
\alias{inpe_station_data}
\title{Get climate data from stations INPE}
\usage{
inpe_station_data(station_id = 31973, start_date = "2005/01/01",
  end_date = "2005/02/02")
}
\arguments{
\item{station_id}{A numeric vector with the station id.}

\item{start_date}{Start date.}

\item{end_date}{End date (Maximum of one year between start and end dates).}
}
\value{
A data frame containing the climate data and attributes.
}
\description{
Get climate data from stations INPE
}

