% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcds.R
\name{inpe_stations}
\alias{inmet_stations}
\alias{inpe_stations}
\title{List all available climate stations}
\usage{
inpe_stations(states = "all")

inmet_stations(username, password, states = "all")
}
\arguments{
\item{states}{A character vector with abbreviated Brazilian states names or "all" for all PCDs.}

\item{username}{INMET username. Only used if you are querying INMET.}

\item{password}{INMET password.}
}
\value{
A data frame.
}
\description{
List all available climate stations
}

