% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_pcds.R
\name{nearest_inpe_station}
\alias{nearest_inpe_station}
\title{Order climate stations (PCDs) by their distances in kilometers from a given coordinate.}
\usage{
nearest_inpe_station(lat = -22.410259, long = -47.560156)
}
\arguments{
\item{lat}{Latitude in decimals}

\item{long}{Longitude in decimals}
}
\value{
A data frame ordered by the distances to the PCDs.
}
\description{
Order climate stations (PCDs) by their distances in kilometers from a given coordinate.
}

