% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_stations.R
\name{nearest_stations}
\alias{nearest_stations}
\title{Order climate stations by their distances in kilometers from a given point.}
\usage{
nearest_stations(lat = -22.410259, long = -47.560156, source = "both")
}
\arguments{
\item{lat}{Latitude in decimals.}

\item{long}{Longitude in decimals.}

\item{source}{"inpe", "inmet", or "both" (default).}
}
\value{
A data frame ordered by the distances to the stations.
}
\description{
Order climate stations by their distances in kilometers from a given point.
}

